Remove non-generic models from the models package
The moved models are better defined in the services that ‘own’ and return the models
