--- src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/model/BookingSummaryForAvailability.kt
@@ -1,11 +1,0 @@
-package uk.gov.justice.digital.hmpps.approvedpremisesapi.model
-
-import java.time.LocalDate
-
-interface BookingSummaryForAvailability {
-  fun getArrivalDate(): LocalDate
-  fun getDepartureDate(): LocalDate
-  fun getArrived(): Boolean
-  fun getIsNotArrived(): Boolean
-  fun getCancelled(): Boolean
-}
