--- src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/model/UserWorkload.kt
@@ -1,7 +1,0 @@
-package uk.gov.justice.digital.hmpps.approvedpremisesapi.model
-
-data class UserWorkload(
-  var numTasksPending: Int,
-  var numTasksCompleted7Days: Int,
-  var numTasksCompleted30Days: Int,
-)
